Move antlr4 output from source into generated-sources
Otherwise maven detects source changes even if there are any and forces
re-compilation.
