+++ .gitignore
@@ -61,7 +61,6 @@ sandbox/crate/data/
 sandbox/crate/repo/
 sandbox/crate2/data/
 sandbox/crate2/repo/
-**/antlr/
 **/gen/
 target/
 *.factorypath
