+++ libs/pgwire/pom.xml
@@ -21,13 +21,8 @@
         <version>4.13.1</version>
         <configuration>
           <sourceDirectory>${basedir}/src/main/antlr</sourceDirectory>
-          <outputDirectory>${basedir}/src/main/java/io/crate/protocols/postgres/antlr/v4</outputDirectory>
           <visitor>true</visitor>
           <listener>false</listener>
-          <arguments>
-            <argument>-package</argument>
-            <argument>io.crate.protocols.postgres.antlr.v4</argument>
-          </arguments>
         </configuration>
         <executions>
           <execution>
