+++ libs/sql-parser/src/main/java/io/crate/sql/parser/AstBuilder.java
@@ -34,32 +34,31 @@ import java.util.RandomAccess;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
-import org.jetbrains.annotations.Nullable;
-
 import org.antlr.v4.runtime.ParserRuleContext;
 import org.antlr.v4.runtime.Token;
 import org.antlr.v4.runtime.tree.ParseTree;
 import org.antlr.v4.runtime.tree.TerminalNode;
+import org.jetbrains.annotations.Nullable;
 
 import io.crate.common.collections.Lists2;
 import io.crate.sql.ExpressionFormatter;
-import io.crate.sql.parser.antlr.v4.SqlBaseLexer;
-import io.crate.sql.parser.antlr.v4.SqlBaseParser;
-import io.crate.sql.parser.antlr.v4.SqlBaseParser.BitStringContext;
-import io.crate.sql.parser.antlr.v4.SqlBaseParser.CloseContext;
-import io.crate.sql.parser.antlr.v4.SqlBaseParser.ConflictTargetContext;
-import io.crate.sql.parser.antlr.v4.SqlBaseParser.DeclareContext;
-import io.crate.sql.parser.antlr.v4.SqlBaseParser.DeclareCursorParamsContext;
-import io.crate.sql.parser.antlr.v4.SqlBaseParser.DirectionContext;
-import io.crate.sql.parser.antlr.v4.SqlBaseParser.DiscardContext;
-import io.crate.sql.parser.antlr.v4.SqlBaseParser.FetchContext;
-import io.crate.sql.parser.antlr.v4.SqlBaseParser.IsolationLevelContext;
-import io.crate.sql.parser.antlr.v4.SqlBaseParser.QueryContext;
-import io.crate.sql.parser.antlr.v4.SqlBaseParser.QueryOptParensContext;
-import io.crate.sql.parser.antlr.v4.SqlBaseParser.SetTransactionContext;
-import io.crate.sql.parser.antlr.v4.SqlBaseParser.StatementsContext;
-import io.crate.sql.parser.antlr.v4.SqlBaseParser.TransactionModeContext;
-import io.crate.sql.parser.antlr.v4.SqlBaseParserBaseVisitor;
+import io.crate.sql.parser.antlr.SqlBaseLexer;
+import io.crate.sql.parser.antlr.SqlBaseParser;
+import io.crate.sql.parser.antlr.SqlBaseParser.BitStringContext;
+import io.crate.sql.parser.antlr.SqlBaseParser.CloseContext;
+import io.crate.sql.parser.antlr.SqlBaseParser.ConflictTargetContext;
+import io.crate.sql.parser.antlr.SqlBaseParser.DeclareContext;
+import io.crate.sql.parser.antlr.SqlBaseParser.DeclareCursorParamsContext;
+import io.crate.sql.parser.antlr.SqlBaseParser.DirectionContext;
+import io.crate.sql.parser.antlr.SqlBaseParser.DiscardContext;
+import io.crate.sql.parser.antlr.SqlBaseParser.FetchContext;
+import io.crate.sql.parser.antlr.SqlBaseParser.IsolationLevelContext;
+import io.crate.sql.parser.antlr.SqlBaseParser.QueryContext;
+import io.crate.sql.parser.antlr.SqlBaseParser.QueryOptParensContext;
+import io.crate.sql.parser.antlr.SqlBaseParser.SetTransactionContext;
+import io.crate.sql.parser.antlr.SqlBaseParser.StatementsContext;
+import io.crate.sql.parser.antlr.SqlBaseParser.TransactionModeContext;
+import io.crate.sql.parser.antlr.SqlBaseParserBaseVisitor;
 import io.crate.sql.tree.AddColumnDefinition;
 import io.crate.sql.tree.AliasedRelation;
 import io.crate.sql.tree.AllColumns;
