+++ server/pom.xml
@@ -56,13 +56,8 @@
         <version>4.13.1</version>
         <configuration>
           <sourceDirectory>${basedir}/src/main/antlr</sourceDirectory>
-          <outputDirectory>${basedir}/src/main/java/io/crate/signatures/antlr</outputDirectory>
           <visitor>true</visitor>
           <listener>false</listener>
-          <arguments>
-            <argument>-package</argument>
-            <argument>io.crate.signatures.antlr</argument>
-          </arguments>
         </configuration>
         <executions>
           <execution>
